test(inapp): fix code rabbit PR comments SDK-5433

+++ clevertap-core/src/main/java/com/clevertap/android/sdk/inapp/delay/InAppDataExtractor.kt
@@ -55,7 +55,7 @@ internal class InActionDataExtractor : InAppDataExtractor<InActionResult> {
     }
     override fun createDiscardedResult(id: String): InActionResult {
         return InActionResult.Discarded(
-            id,
+            id.toLong(),
             "Timer expired while app was backgrounded"
         )
     }
